+++ aftiktuna-rust/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Crew size limit when recruiting has been raised from 2 to 3
 - Recruitable characters and crew members in json definitions can now be set to use a random character profile
 - Aftik corpses can now be set to use a color from a random character profile
+- The "recruitable" symbol type (for location defintions) has now been replaced by a more general "character" symbol type
 - Loot tables are no longer hardcoded, and can be defined in `assets/loot_table/`
 - Characters may now push crew members that are in the way for certain actions
 - The tent in the eyesaur forest is now a searchable container